+++ src/by/it/dorostchenok/matlab/ConsoleRunner.java
@@ -1,13 +1,13 @@
-package by.it.dorostchenok.jd01_09;
+package by.it.dorostchenok.matlab;
 
-import by.it.dorostchenok.jd01_09.exception.BadOperationException;
-import by.it.dorostchenok.jd01_09.exception.BadValueException;
-import by.it.dorostchenok.jd01_09.exception.VariableNotDefinedException;
-import by.it.dorostchenok.jd01_09.interfaces.VariableStorageDAO;
-import by.it.dorostchenok.jd01_09.services.Parser;
-import by.it.dorostchenok.jd01_09.services.VariableStorageMapImpl;
-import by.it.dorostchenok.jd01_09.values.Value;
-import by.it.dorostchenok.jd01_09.values.Variable;
+import by.it.dorostchenok.matlab.exception.BadOperationException;
+import by.it.dorostchenok.matlab.exception.BadValueException;
+import by.it.dorostchenok.matlab.exception.VariableNotDefinedException;
+import by.it.dorostchenok.matlab.interfaces.VariableStorageDAO;
+import by.it.dorostchenok.matlab.services.Parser;
+import by.it.dorostchenok.matlab.services.VariableStorageMapImpl;
+import by.it.dorostchenok.matlab.values.Value;
+import by.it.dorostchenok.matlab.values.Variable;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -15,7 +15,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.Collections;
 import java.util.List;
 
-import by.it.dorostchenok.jd01_09.services.VariableComparator;
+import by.it.dorostchenok.matlab.services.VariableComparator;
 
 public class ConsoleRunner {
