+++ src/by/it/dorostchenok/matlab/services/VariableComparator.java
@@ -1,6 +1,6 @@
-package by.it.dorostchenok.jd01_09.services;
+package by.it.dorostchenok.matlab.services;
 
-import by.it.dorostchenok.jd01_09.values.Value;
+import by.it.dorostchenok.matlab.values.Value;
 
 
 import java.util.Comparator;
